requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
django-extensions==3.2.3
openpyxl==3.1.2
reportlab==4.0.4
django-allauth==0.57.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
dj-rest-auth==7.0.1
